Rename files to use UpperCamelCaseCluster/Logic for general-commissioning-server
