+++ src/app/clusters/network-commissioning/CodegenInstance.cpp
@@ -17,7 +17,7 @@
 #include "CodegenInstance.h"
 
 #include <app/clusters/general-commissioning-server/CodegenIntegration.h>
-#include <app/clusters/general-commissioning-server/general-commissioning-cluster.h>
+#include <app/clusters/general-commissioning-server/GeneralCommissioningCluster.h>
 #include <app/clusters/network-commissioning/NetworkCommissioningCluster.h>
 #include <data-model-providers/codegen/CodegenDataModelProvider.h>
